// SISPRO/yitpro-react/src/model/FlujoPago.model.ts
export interface FlujoPagoModel {
    idFlujoPago: number;
    idProyecto: number;
    precioHora: number;
    horasTotales: number;
    horasAmortizar: number;
    porcIVA: number;
    fechaCreo: string;
    idUCreo: number;
    fechaMod: string | null;
    idUMod: number | null;
    activo: boolean;
    ene: number;
    feb: number;
    mar: number;
    abr: number;
    may: number;
    jun: number;
    jul: number;
    ago: number;
    sep: number;
    oct: number;
    nov: number;
    dic: number;
    totalFacturado: number | null;
    totalPagado: number | null;
    saldo: number | null;
    totalFacturadoHoras: number | null;
    totalPagadoHoras: number | null;
    saldoHoras: number | null;
    claveProy: string;
    nombreProy: string;
    lider: string;
    cliente: string;
    flujoDetalle: FlujoPagoDetModel[];
    totalProyecto: number | null;
    totalProyectoHoras: number | null;
    totalAmortizadas: number | null;
    totalAmortizadoHoras: number | null;
    saldoAmortizarHoras: number | null;
    saldoAmortizar: number | null;
    pendienteAmortizar: number;
    amortizadas: number;
    facturado: number;
    pendienteFacturar: number;
    avance: number;
}
export interface FlujoPagoDetModel {
    idFlujoPagoDet: number;
    idFlujoPago: number;
    secuencia: number | null;
    concepto: string;
    horas: number | null;
    amortizadas: number | null;
    procentaje: number | null;
    monto: number | null;
    total: number | null;
    iVA: number | null;
    fechaDevOriginal: string | null;
    fechaDev: string | null;
    fechaFactura: string | null;
    fechaProgramadaPago: string | null;
    fechaPagoReal: string | null;
    facturable: boolean;
    facturada: boolean;
    pagada: boolean;
    vencido: boolean;
    tipoFecha: number;
    fecha: string | null;
    comentarios: string;
    factura: string;
    proyecto: string;
    idProyecto: number;
}